This schema is a first cut at some Cobol from daffodil.
Updated to use 'V' virtual decimal point.

Uses zoned decimal with trailing overpunched sign for
one element.

Requires Daffodil 3.5.0 which is where this 'V' feature
was first implemented. (Or a snapshot build with the feature.)

Orignally, this example had a date as a packed decimal.

Per issue #4, date as packed is unusual for Cobol.

Changed to string as this is more standard and more portable.