Version 0.0.2-SNAPSHOT: Fixed leading sign on zoned error.
The COBOL says SIGN LEADING On the TOTAL-CHARGES field.
It was using default trailing sign. Changed this and the test data.
This was not only leading/trailing inverted, but it had
'y' as the expected sign character, which is the asciiStandard convention
for -9 digit. This schema and test data is ebcdic however,
so that character should be 'R'.

That lead to discovery of DAFFODIL-2873 - that zoned
ebcdic overpunched signs do not work.

Hence, this example now requires daffodil version 3.7.0 or higher,
(or a 3.7.0-SNAPSHOT build with the fix for DAFFODIL-2873 which was commited
on 2024-01-21)

Added example of using xs:date instead of treating date as strings.
COBOL has no date data type, but it is natural to want to use DFDL
date/time types so the parsed output XML/JSON/Other creates a date type.

Added comments about the COBOL types and PIC clauses as guidance for
COBOL users who encounter such combinations and to assist with Issue #3
(https://github.com/DFDLSchemas/Cobol/issues/3).

Test is two-pass round trip since F9 is canonicalized to C9
when unparsing. Retained this in the test data just to verify
that this canonicalization on unparsing is happening.

Update sbt to latest.

Update Daffodil to 3.7.0-SNAPSHOT as this is required for the fix to
DAFFODIL-2873 (fixes zoned ebcdic overpunched signs).

https://github.com/DFDLSchemas/Cobol/issues/7